converted file from js to tsx and make the changes
